Imported hand gesture to camera feed
